Added new functions, fixed some errors
Added ability for matricies to transform a point. Changed vector
multiplication operators to be based on floats

+++ Source/Vector/Vector/Vector2.cpp
@@ -111,28 +111,28 @@ Vector2& Vector2::operator-=(const Vector2 &rhs)
 	return *this;
 }
 
-Vector2 Vector2::operator*(const int &rhs)
+Vector2 Vector2::operator*(const float &rhs)
 {
 	Vector2 vector = *this;
 	vector *= rhs;
 	return vector;
 }
 
-Vector2& Vector2::operator*=(const int &rhs)
+Vector2& Vector2::operator*=(const float &rhs)
 {
 	mX *= rhs;
 	mY *= rhs;
 	return *this;
 }
 
-Vector2 Vector2::operator/(const int &rhs)
+Vector2 Vector2::operator/(const float &rhs)
 {
 	Vector2 vector = *this;
 	vector /= rhs;
 	return vector;
 }
 
-Vector2& Vector2::operator/=(const int &rhs)
+Vector2& Vector2::operator/=(const float &rhs)
 {
 	mX /= rhs;
 	mY /= rhs;
