+++ Source/Vector/Vector/Vector2.h
@@ -29,10 +29,10 @@ public:
 	Vector2& operator+=(const Vector2 &rhs);
 	Vector2 operator-(const Vector2 &rhs);
 	Vector2& operator-=(const Vector2 &rhs);
-	Vector2 operator*(const int &rhs);
-	Vector2& operator*=(const int &rhs);
-	Vector2 operator/(const int &rhs);
-	Vector2& operator/=(const int &rhs);
+	Vector2 operator*(const float &rhs);
+	Vector2& operator*=(const float &rhs);
+	Vector2 operator/(const float &rhs);
+	Vector2& operator/=(const float &rhs);
 
 private:
 	float mX;
