+++ Source/Vector/Vector/Vector3.h
@@ -32,10 +32,11 @@ public:
 	Vector3& operator+=(const Vector3 &rhs);
 	Vector3 operator-(const Vector3 &rhs);
 	Vector3& operator-=(const Vector3 &rhs);
-	Vector3 operator*(const int &rhs);
-	Vector3& operator*=(const int &rhs);
-	Vector3 operator/(const int &rhs);
-	Vector3& operator/=(const int &rhs);
+	Vector3 operator*(const float &rhs);
+	Vector3& operator*=(const float &rhs);
+	Vector3 operator/(const float &rhs);
+	Vector3& operator/=(const float &rhs);
+
 
 private:
